refactor(strategies): divided in 1 file per function, renamed files to index and deploy, and created a sub-folder in straetgies containing all of the recursive-farming strat
